Retrieval-Augmented-Generation/04_Rag_Application_with_FAISS/pyproject.toml
[project]
name = "04-rag-application-with-faiss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.25",
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
    "rapidocr-onnxruntime>=1.4.4",
    "tiktoken>=0.9.0",
]
